+++ app/[event]/components/filter.tsx
@@ -2,8 +2,10 @@
 
 import { useDebouncedCallback } from 'use-debounce';
 
-import { FactionSelect, SearchField } from 'components';
-import { useFilter } from './filter-context';
+import { SearchField, Select } from 'components';
+import { getAllFactions } from 'lib/get-value';
+
+import { FactionOptions, useFilter } from './filter-context';
 
 export const Filter = () => {
   const filter = useFilter();
@@ -18,7 +20,19 @@ export const Filter = () => {
         defaultValue={filter.query}
         onChange={e => debounce(e.target.value)}
       />
-      <FactionSelect value={filter.faction} onChange={filter.setFaction} />
+      <Select
+        size="small"
+        aria-label="Faction"
+        value={filter.faction}
+        onChange={e => filter.setFaction(e.target.value as FactionOptions)}
+      >
+        <Select.Option value="all">All Factions</Select.Option>
+        {getAllFactions().map(({ id, name }) => (
+          <Select.Option key={id} value={id}>
+            {name}
+          </Select.Option>
+        ))}
+      </Select>
     </div>
   );
 };
