+++ components/index.ts
@@ -3,7 +3,6 @@ export * from './caption';
 export * from './card';
 export * from './center';
 export * from './container';
-export * from './faction-select';
 export * from './input';
 export * from './label';
 export * from './link';
